Update LoginGUI.java
I, for the love of god, could not get this updated LoginGUI to run properly, so I made some minor changes and additions without touching it too much. Hopefully they work as intended.

Made a new Patient and CareGiver object in this class. These public variables carry over into the Patient/Caregiver GUI and store their data so we can access it there. Switched the user.add("whatever") to patient, since we aren't making User objects on its own. Added the setter methods for the careGiver to match with the Patient sign up. Made a ReportList object in case we need it as well.